feat: mailing lists ui + refactor components
