+++ ui/app/liste-diffusion/page.tsx
@@ -1,31 +1,15 @@
 "use client";
-import {
-  Box,
-  Button,
-  FormControl,
-  FormErrorMessage,
-  FormLabel,
-  Heading,
-  HStack,
-  Select,
-  useToast,
-} from "@chakra-ui/react";
+import { Heading } from "@chakra-ui/react";
 import { useQuery } from "@tanstack/react-query";
-import { useForm } from "react-hook-form";
 import { IJob, JOB_STATUS_LIST } from "shared/models/job.model";
-import { IReqGetMailingList } from "shared/routes/mailingList.routes";
-import { DOCUMENT_TYPES } from "shared/routes/upload.routes";
 
-import Table from "../../components/table/Table";
-import { Bin } from "../../theme/icons/Bin";
-import { DownloadLine } from "../../theme/icons/DownloadLine";
 import { api } from "../../utils/api.utils";
-import { formatDate } from "../../utils/date.utils";
 import Breadcrumb, { PAGES } from "../components/breadcrumb/Breadcrumb";
+import Form from "./components/Form";
+import GeneratingMailingList from "./components/GeneratingMailingList";
+import ListMailingList from "./components/ListMailingList";
 
 const ListeDiffusionPage = () => {
-  const toast = useToast();
-
   const { data: mailingLists, refetch } = useQuery<IJob[]>({
     queryKey: ["mailingLists"],
     queryFn: async () => {
@@ -35,30 +19,17 @@ const ListeDiffusionPage = () => {
     },
   });
 
-  const {
-    handleSubmit,
-    formState: { errors, isSubmitting },
-    register,
-  } = useForm<IReqGetMailingList>();
-
-  const onSubmit = async (data: IReqGetMailingList) => {
-    await api.post("/mailing-list", { source: data.source });
+  const handleGenerate = async () => {
     await refetch();
-
-    toast({
-      title:
-        "La liste de diffusion est en cours de génération. Vous pouvez revenir sur cette page dans quelques minutes.",
-      status: "success",
-      duration: 10000,
-      isClosable: true,
-    });
   };
 
-  const onDeleteMailingList = async (mailingList_id: string) => {
-    await api.delete(`/mailing-list/${mailingList_id}`);
-    if (typeof window !== "undefined" && window.location)
-      window.location.reload();
-  };
+  const generatingMailingList = mailingLists?.find((mailingList) =>
+    [
+      JOB_STATUS_LIST.RUNNING,
+      JOB_STATUS_LIST.PENDING,
+      JOB_STATUS_LIST.WILLSTART,
+    ].includes(mailingList.status as JOB_STATUS_LIST)
+  );
 
   return (
     <>
@@ -67,132 +38,13 @@ const ListeDiffusionPage = () => {
         Liste de diffusion
       </Heading>
 
-      <Box w={{ base: "100%", md: "50%" }}>
-        <form onSubmit={handleSubmit(onSubmit)}>
-          <Box>
-            <FormControl
-              isInvalid={!!errors.source}
-              mb={5}
-              isDisabled={isSubmitting}
-            >
-              <FormLabel>Source</FormLabel>
-              <Select
-                isInvalid={!!errors.source}
-                placeholder="Choisir la source"
-                {...register("source", {
-                  required: "Obligatoire: Vous devez choisir la source",
-                  validate: (value) => {
-                    return (
-                      value && Object.values(DOCUMENT_TYPES).includes(value)
-                    );
-                  },
-                })}
-              >
-                {[
-                  DOCUMENT_TYPES.VOEUX_AFFELNET_MAI_2023,
-                  DOCUMENT_TYPES.VOEUX_AFFELNET_JUIN_2023,
-                ].map((type) => (
-                  <option key={type}>{type}</option>
-                ))}
-              </Select>
-              <FormErrorMessage>{errors.source?.message}</FormErrorMessage>
-            </FormControl>
-          </Box>
+      {generatingMailingList ? (
+        <GeneratingMailingList mailingList={generatingMailingList} />
+      ) : (
+        <Form onSuccess={handleGenerate} />
+      )}
 
-          <HStack spacing="4w" mt={8}>
-            <Button variant="primary" type="submit" isLoading={isSubmitting}>
-              Générer la liste
-            </Button>
-          </HStack>
-        </form>
-      </Box>
-
-      <Box>
-        <Table
-          mt={4}
-          data={mailingLists || []}
-          columns={{
-            source: {
-              id: "source",
-              size: 100,
-              header: () => "Source",
-              cell: ({ row }) => row.original.payload?.source,
-            },
-            status: {
-              id: "status",
-              size: 100,
-              header: () => "Statut",
-              cell: ({ row }) => {
-                return (
-                  <>
-                    {
-                      {
-                        pending: "En attente",
-                        will_start: "Programmé",
-                        running: "En cours",
-                        finished: "Terminé",
-                        blocked: "Bloqué",
-                        errored: "Erreur",
-                      }[row.original.status]
-                    }
-                    {row.original.status === JOB_STATUS_LIST.RUNNING &&
-                      row.original.payload?.processed && (
-                        <> ({row.original.payload?.processed} %)</>
-                      )}
-                  </>
-                );
-              },
-            },
-            date: {
-              id: "date",
-              size: 100,
-              header: () => "Date de génération",
-              cell: ({ row }) => {
-                return (
-                  row.original.created_at &&
-                  formatDate(
-                    //@ts-ignore
-                    row.original.created_at as string,
-                    "dd/MM/yyyy à HH:mm"
-                  )
-                );
-              },
-            },
-
-            actions: {
-              id: "actions",
-              size: 25,
-              header: () => "Actions",
-              cell: ({ row }) => {
-                if (row.original.status !== JOB_STATUS_LIST.FINISHED) {
-                  return null;
-                }
-
-                return (
-                  <HStack spacing={4}>
-                    <a
-                      href={`/api/mailing-lists/${row.original._id}/download`}
-                      title="Télécharger le fichier"
-                      target="_blank"
-                      rel="noopener noreferrer"
-                    >
-                      <DownloadLine w="1w" />
-                    </a>
-                    <Bin
-                      boxSize="5"
-                      color="red_marianne"
-                      cursor="pointer"
-                      onClick={() =>
-                        onDeleteMailingList(row.original._id.toString())
-                      }
-                    />
-                  </HStack>
-                );
-              },
-            },
-          }}
-        />
-      </Box>
+      <ListMailingList mailingLists={mailingLists} onDelete={refetch} />
     </>
   );
 };
